+++ NEWS.md
@@ -723,6 +723,10 @@ Library improvements
     optional keyword argument, as in `reduce(op, itr; init=v0)`. Similarly for `foldl`,
     `foldr`, `mapreduce`, `mapfoldl` and `mapfoldr`. ([#27711])
 
+  * `Sys.glibc()` determines whether glibc is linked to Julia, and `Sys.glibc_version()`
+    returns the version of the linked glibc if applicable. ([#27223])
+
+
 Compiler/Runtime improvements
 -----------------------------
 
@@ -1640,5 +1644,6 @@ Command-line option changes
 [#27164]: https://github.com/JuliaLang/julia/issues/27164
 [#27189]: https://github.com/JuliaLang/julia/issues/27189
 [#27212]: https://github.com/JuliaLang/julia/issues/27212
+[#27223]: https://github.com/JuliaLang/julia/issues/27223
 [#27248]: https://github.com/JuliaLang/julia/issues/27248
 [#27401]: https://github.com/JuliaLang/julia/issues/27401